Exception added to open() and fixed \r\n
The DelimitedInputFormat.open() method did not check if it has parsers
registered. Added throwing an exception when no parser is registered
when calling open. CsvInputFormat does now accept windows line endings
with standards settings. The |\r\n" will be replaced by "\n_"( _ is a
space)... To prevent from errors the returned length is set to the
space. So that this byte is known as parsed and will not get interpreted
anymore.